feat: implement load default image

+++ src/routes/words.ts
@@ -1,8 +1,6 @@
 import { Router } from 'express';
 import { loader } from '../middleware/files';
-import {
-  deleteWord, getAllWords, insertWord, updateWord,
-} from '../storage/words';
+import { deleteWord, getAllWords, insertWord, updateWord } from '../storage/words';
 import { TFiles, TWord } from '../types';
 
 const fs = require('fs/promises');
@@ -49,9 +47,15 @@ router.post('/create', loader.fields([{ name: 'img' }, { name: 'audio' }]), asyn
 
     wordData.audioSRC = audioPath.secure_url;
     fs.unlink((req.files as TFiles).audio[0].path);
-    const imgPath = await cloudinary.uploader.upload((<TFiles>req.files).img[0].path);
-    wordData.imageSRC = imgPath.secure_url;
-    fs.unlink((req.files as TFiles).img[0].path);
+
+    if ((<TFiles>req.files).img) {
+      const imgPath = await cloudinary.uploader.upload((<TFiles>req.files).img[0].path);
+      wordData.imageSRC = imgPath.secure_url;
+      fs.unlink((req.files as TFiles).img[0].path);
+    } else {
+      wordData.imageSRC =
+        'https://res.cloudinary.com/dshgus1qp/image/upload/v1626286092/english-for-kids/img/default.jpg';
+    }
   } catch (e) {
     console.log(e, 'something went wrong');
   }
